Move the example and component readers into _modules
The /example and /example-ssr JSON endpoints were copies of each other.
Both now call getExampleContent.js, and the component endpoint calls
getComponentContent.js. The modules export the ExampleContent and
ComponentContent typedefs, and the load functions cast the fetched JSON to
them, so the pages are type-checked instead of seeing `any`. The root
layout and guide loads type their sections the same way.

The loads pass a missing example or component through as a 404 instead of
a 500. The components endpoint no longer looks for src/content/components,
which doesn't exist. The Playground link opens on the Svelte version the
site is built with.
